refactor(testing): разделён RSSReaderTestSupport
монолитный RSSReaderTestSupport заменён focused test support файлами
SwiftData harness, scripted HTTP client, DTO factories, logging doubles и source-management fixtures теперь живут отдельно
