+++ README.md
@@ -532,7 +532,7 @@
 - [x] `Feed Parsing Module Split`: разделить `FeedParserService` на XML document builder, feed kind detection, RSS parser, Atom parser и parse pipeline diagnostics, чтобы будущие tests для parsing / normalizer не зависели от одного большого файла с несколькими уровнями ответственности;
 - [x] `Refresh Pipeline Boundary Cleanup`: проверить `FeedRefreshService` после накопленных archive / retention / diagnostics изменений и выделить из него отдельные шаги conditional request, not-modified handling, fetched response reconciliation, diagnostics/log persistence и batch orchestration, не меняя публичный `FeedRefreshCoordinating` contract без необходимости;
 - [x] `Loading Indicator Contract Audit`: пройти оставшиеся loading states после `Shared Loading Indicator Tint And Reuse` и явно зафиксировать, что runtime UI не использует системный `ProgressView` / UIKit indicator, а loading states должны использовать `AppRefreshIndicator`, чтобы UI не возвращался к смешанным indicator semantics перед UI tests;
-- [ ] `Test Harness Decomposition`: разложить `RSSReaderTestSupport` на focused builders / fixtures: SwiftData harness, scripted HTTP client, DTO factories, logging doubles и source-management fixtures; тесты должны импортировать общие helpers без неявной зависимости от feed refresh harness;
+- [x] `Test Harness Decomposition`: разложить `RSSReaderTestSupport` на focused builders / fixtures: SwiftData harness, scripted HTTP client, DTO factories, logging doubles и source-management fixtures; тесты должны импортировать общие helpers без неявной зависимости от feed refresh harness;
 - [ ] `App Dependencies Test Suite Split`: разделить большой `AppDependenciesTests` на suites по composition areas: SwiftData configuration, sync bootstrap policy, runtime orchestrator wiring, background refresh dependencies, cache directories и app action entry points; это должно подготовить инфраструктуру к эпику `Testing` и снизить стоимость targeted test runs;
 - [ ] `Roadmap TODO Cleanup`: перенести временные inline TODO для будущих OPML import/export flows из runtime UI-кода в roadmap / Post-MVP backlog или заменить их нейтральными disabled actions, чтобы текущий код не хранил незакрытые комментарии по отложенной функциональности.
