+++ RSSReaderTests/TestSupport/SourceManagementFixtures.swift
@@ -1,0 +1,14 @@
+import Foundation
+@testable import RSSReader
+
+extension SourceManagementScreenState {
+    static func makePreviewFixture(
+        presentedScenarioID: SourceManagementScenarioID? = nil
+    ) -> SourceManagementScreenState {
+        var state = SourceManagementScreenState()
+        if let presentedScenarioID {
+            state.presentScenario(presentedScenarioID)
+        }
+        return state
+    }
+}
